Handle FILE: keytab with a colon in the path
If we are given a keytab name beginning with FILE:, then treat the rest as
a path, even if it happens to contain another colon.  Without this, a keytab
like FILE:/tkt/foo:bar will be passed to kt_resolve as /tkt/foo:bar, which
at a minimum means it doesn't get the WRFILE treatment, and may in fact do
completely the wrong thing.