finalize: initialize krb5_key_data before use
When using kadm5_chpass_principal_with_key(), we pass down an array of
krb5_key_data, each of which contains one of the new keys.  However,
these structures also contain salt information which is sent over the
wire, including a salt string.  If we don't zero the structure before
starting, the salt pointer and length are garbage, and we crash -- even
though they are not actually going to be used on the server!