Configure switch for default service principal
Add a configure switch that sets the default service principal used by
clients when connecting to the server.  This avoids the need for every
user of these clients to remember and use the -P switch.

Even when a default is set, the original behavior of using an appropriate
host principal can be obtained by using the -P switch to explicitly give
a principal name of "-" or the empty string.