server: set a 30s SQLite busy timeout
When obtaining or upgrading a database lock, SQlite doesn't block until
the lock is available; instead, if the lock cannot be obtained, it simply
fails with SQLITE_BUSY and expects the caller to try again later.

This can be mediated by setting a busy handler or timeout, which gives the
application an opportunity to tell SQLite to retry the lock, presumably
after some delay.  In our case, we set a timeout of 30 seconds.

The timeout mechanism works by setting a busy handler which sleeps for a
short time (on the order of milliseconds) and then retries, until the total
amount of time spent sleeping is greater than the configured timeout.  This
method is unfortunately not fair -- it does not guarantee an upper bound on
how long a thread will have to wait until it can have the lock.  Further,
since the default busy handler backs off exponentially, a thread that has
been waiting a while actually have a lower chance of obtaining the lock than
does a thread that has just started waiting.

We try to mitigate this failing by setting a very long timeout (30s), on the
theory that waiting a long time is better than failing, and thus a timeout
should be an exceptionally rare event.  Unfortunately, this may still not be
enough -- SQLite's locking heirarchy includes some possibility of deadlock,
in which case one of the affected threads will get SQLITE_BUSY anyway.