server: hold a lock while using the database
SQLite doesn't actually provide a very useful concurrency model.  It
ensures the database doesn't get corrupted, but at the expense of any
operation being able to fail with a "database is busy" error, requiring
that the application handle retrying the operation.  This is true even
if a busy handler or timeout is set, allowing locking operations to be
retried, both because the lock retry is unfair and because SQLite's lock
heirarchy contains deadlocks which can be resolved only by causing some
thread to fail.

To avoid this whole situation, we obtain an exclusive file-level lock
on a lock file before opening the database, ensuring that no other server
process is attempting to do the same.