Drop the -a switch
Remove the -a switch, which was used to specify the file or LDAP group
name used by the configured admin method. The ACL file is now fixed, as
is the group used by the old-style LDAP group method. The group used by
the ldapgroups-std method can be configured by means of a new appdefaults
setting.