Require (and document) certain LDAP options
when the admin ACL method is ldapgroups-std, require the LDAP URI,
base DN, and group DN to be specified in /etc/krb5.conf. These
settings no longer have built-in defaults that are specific to CMU
or a particular environment.

Additionally, document the options that can be used to configure
the ldapgroups-std method.