systemd unit: don't consider closed connection a failure
If a client shuts down the SSL connection in a normal fashion, rekeysrv
will exit with status 1, which systemd counts as a failure. This causes
the "failed" unit to be kept around forever, eventually making systemd
quite upset, especially in a realm like CS.CMU.EDU with thousands of
client connections every day.

Avoid treating exit status 1 as a failure.