Add the daily cron job
For andy3 server builds (sub-krb_rekey_server), the file will be
present but not instlled, due to a directive in version_build_config.

For misc collection builds, this will be present but not installed,
since it is not mentioned in the depot.conf and /etc/ is not under
depot's control anyway.